RestaurantsMapActivity moved to *views package
