+++ app/src/main/java/com/tolmachevroman/restaurants/views/RestaurantsMapActivity.kt
@@ -1,14 +1,14 @@
-package com.tolmachevroman.restaurants
+package com.tolmachevroman.restaurants.views
 
-import android.support.v7.app.AppCompatActivity
 import android.os.Bundle
-
+import android.support.v7.app.AppCompatActivity
 import com.google.android.gms.maps.CameraUpdateFactory
 import com.google.android.gms.maps.GoogleMap
 import com.google.android.gms.maps.OnMapReadyCallback
 import com.google.android.gms.maps.SupportMapFragment
 import com.google.android.gms.maps.model.LatLng
 import com.google.android.gms.maps.model.MarkerOptions
+import com.tolmachevroman.restaurants.R
 
 class RestaurantsMapActivity : AppCompatActivity(), OnMapReadyCallback {
