+++ app/src/main/res/layout/activity_restaurants_map.xml
@@ -5,4 +5,4 @@
     android:name="com.google.android.gms.maps.SupportMapFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.tolmachevroman.restaurants.RestaurantsMapActivity" />
+    tools:context="com.tolmachevroman.restaurants.views.RestaurantsMapActivity" />
